Fix segfault under BSDs
Passing a NULL or the string "any" under Linux requests pcap_open_live()
open all devices. On BSDs, this behaviour is not supported, so
pcap_open_live() will fail, then segfault when pcap_geterr() tries to
print out NULL (on Linux, the error is "(null): error message").

If no device is specified, look up the default device. For Linux, the
keyword "any" can be used instead of a device name.

reported-by: Juraj Hlista