Begin receiving packets upon opening the device
Simplify the interface by immediately receiving packets when the device
is opened. By default, all packets are sniffed. A filter can be applied
when the device is opened or changed later using filter/3.

Remove the loop/1 function. This also prevents multiple threads being
spawned to handle the same pcap descriptor.