Optionally disable dirty schedule support
Allow ewpcap to compile on versions of erlang pre-20 or on builds with
dirty scheduler support explictily disabled at compile time.

Usage of the dirty scheduler is disabled by setting a compile time flag.
An improvement would be to use a rebar.config.script to set the flag
based on:

* an environment variable
* otherwise, by checking erlang:system_info(dirty_cpu_schedulers)

Another way would be check on NIF load using enif_system_info().