lookupdev/0,findalldevs/0: use dirty scheduler
pcap_lookupdev() retrieves a list of the system interfaces using
pcap_findalldevs(), removes the loopback devices then returns the most
appropriate device for use. By default, ewpcap will call
pcap_lookupdev/0 if no device is specified by the caller.

On most systems, pcap_lookupdev() returns quickly. If the system has a
large number of non-loopback interfaces, it is possible it may run longer
than the recommended 1ms and block a scheduler. Note pcap_lookupdev()
should not be in the hot path (it will be called once) and is easy to
work around: specify the device instead of doing a lookup.

Use the dirty scheduler for these functions if available. It would
be possible to create non-blocking versions of these functions (e.g.,
on linux, using the file module and /proc; on bsd, successive calls to
ioctl()) but the point of using pcap is a portable interface.