Use ERL_NIF_DIRTY_JOB_CPU_BOUND for blocking ops
See this thread:

htttp://erlang.org/pipermail/erlang-questions/2018-October/096517.html

The blocking calls are not related to disk operations.

libpcap blocks in poll()/select() waiting for activity from the file
descriptor. The NIF thread running pcap is blocked until the poll()
returns.

Since the underlying pcap descriptor is a socket (Linux) or a file
descriptor (BSD), a sneaky way of forcing libpcap to return early might
be to simply close(2) the fd.